rust-analyzer: refactor notification handlers
